Added the JavaScript library.
This is the first version of the CustomHTMLScroll library.

To use this library, embed it in your code and follow the next steps.

1) Set the height of the HTML element that you want to add the scrollbar
to. 

2) call the "addCustomHTMLScroll" method on the same element.

The library will wrap the element in another block element which will be
called "scrollPane". It adds a scrollbar to the right of the block
element.